Use `loadIntoMemory`
